+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 Version 106:
 
 * Dynamic buffer input areas are mutable
+* Add flat_static_buffer::reset
 
 --------------------------------------------------------------------------------